DeltaTime関数の修正

+++ MyGameEngine/Engine/GameObject.h
@@ -67,5 +67,5 @@ public:
 		return p;
 	}
 
-	int DeltaTime();
+	double DeltaTime();
 };
